Added PSData, which parses SPF strings
A PSData object recieves a string, that is coded in SPF format. SPF is
short for Standard Puzzle Format, a file format to save puzzles.
Basically it is a postscript dictionary, that can be embedded in an
postscript file, that contains no other dictionarys. More information
on SPF format will be added soon.

In method loadFile the loaded file is used to create a PSData object.
If the object is valid, the puzzle type is extracted from that object.
The type will be used to dynamically load the puzzleviewer for this
type.