Changed PuzzleViewer to be a sibling of Display
All PuzzleViewers (yet only KropkiViewer) are now derived from
Display, which reduces the paint-routine in PuzzleApplet to a single
loop.

There happend to be a variable name clash in Kropki-Viewer top, left,
width and height was already used there for different purpose. While
width and height could be changed to local variables to avoid this
conflict top and left had to be renamed to deltaX and deltaY.

The way object inheritance works has also changed. I'm not sure what
is best, but I found a tutorial which convineced me to do it the way
it is done now until I find out more about javascript.