Added support for other languages
Languages, other than german, can be added in the folder language in the
way en.js shows. German is added to the translate.js file and serves
as a backup.

German as backup has been choosen, because not all english
translations do yet exist and also because people are allready used to
the german texts. When all english translations have been done it
might be possible to change the default behaviour.